Cleaned up console commands
