+++ src/main/java/dev/starless/maggiordomo/console/ConsoleCommand.java
@@ -1,0 +1,10 @@
+package dev.starless.maggiordomo.console;
+
+import java.util.List;
+
+public interface ConsoleCommand {
+
+    void execute(String[] args);
+
+    List<String> aliases();
+}
